run/requirements.txt
ultralytics 
pyyaml 
transformers 
opencv-python-headless 
seaborn 
timm 
numpy==1.26.4 
scikit-learn==1.3.2 
huggingface_hub
fastapi
uvicorn
opencv-python
torch
torchvision
pillow
python-multipart